Update library versions in build configuration
Updated the versions of 'fantomas', 'FSharp.Core', 'FsToolkit.ErrorHandling', 'NUnit', and 'NUnit.Analyzers' in the '.config/dotnet-tools.json' and 'paket.lock' files. The updates are necessary to take advantage of new features and fixes provided in the newer versions of these libraries. It is part of our routine maintenance to keep the project up to date with the latest stable versions.